INSUREZ PROJECT/backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==1.4.53
python-dotenv==1.0.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
psycopg2-binary==2.9.9
bcrypt==4.1.2
celery==5.3.4
redis==5.0.1
fastapi-mail==1.4.1